Multi-target fix (#322)
* Support multi target regression for composer

+++ fedot/core/composer/cache.py
@@ -1,3 +1,4 @@
+import glob
 import os
 import shelve
 import uuid
@@ -11,7 +12,7 @@ CachedState = namedtuple('CachedState', 'operation')
 class OperationsCache:
     def __init__(self, db_path=None, clear_exiting=True):
         if not db_path:
-            self.db_path = f'{str(default_fedot_data_dir())}/{str(uuid.uuid4())}'
+            self.db_path = f'{str(default_fedot_data_dir())}/tmp_{str(uuid.uuid4())}'
         else:
             self.db_path = db_path
 
@@ -28,10 +29,13 @@ class OperationsCache:
             _save_cache_for_node(self.db_path, node.descriptive_id,
                                  CachedState(node.fitted_operation))
 
-    def clear(self):
-        for ext in ['bak', 'dir', 'dat']:
-            if os.path.exists(f'{self.db_path}.{ext}'):
-                os.remove(f'{self.db_path}.{ext}')
+    def clear(self, tmp_only=False):
+        if not tmp_only:
+            for ext in ['bak', 'dir', 'dat']:
+                if os.path.exists(f'{self.db_path}.{ext}'):
+                    os.remove(f'{self.db_path}.{ext}')
+        folder_path = f'{str(default_fedot_data_dir())}/tmp_*'
+        clear_folder(folder_path)
 
     def get(self, node):
         found_operation = _load_cache_for_node(self.db_path, node.descriptive_id)
@@ -52,3 +56,10 @@ def _load_cache_for_node(db_path: str, structural_id: str):
         cached_state = cache.get(structural_id, None)
 
     return cached_state
+
+
+def clear_folder(folder_path: str):
+    """ Delete files from chosen folder """
+    temp_files = glob.glob(folder_path)
+    for file in temp_files:
+        os.remove(file)
